Implement a --localize-symbols inplace equivalent
Instead of parsing and rewriting the whole object,
we parse only what we need, modify the buffer in-place in memory,
and write a copy with just our changes applied.

This means we don't need to support all the features of the
ELF standard, just those that deal with symbol tables.
It also means we don't support those obscure features that do
reference the symbol table, because I only implemented
the common stuff, but that's okay in practice.